+++ module-server/src/main/kotlin/sk/momosilabs/recharger/server/service/charging/model/Charging.kt
@@ -6,8 +6,8 @@ import java.time.ZonedDateTime
 data class Charging(
     val time: ZonedDateTime,
     val mileage: Int,
-    val percentageFrom: BigDecimal,
-    val percentageTo: BigDecimal,
+    val percentageFrom: Int,
+    val percentageTo: Int,
     val price: BigDecimal?,
     val kWh: BigDecimal?,
     val provider: ProviderSimple?,
